Update: Make startPackageInstaller an extension of Activity. Add: Clean up cached apk after install
