+++ src/main/kotlin/nya/kitsunyan/foxydroid/screen/ScreenActivity.kt
@@ -2,7 +2,6 @@ package nya.kitsunyan.foxydroid.screen
 
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.os.Parcel
 import android.view.View
@@ -12,17 +11,14 @@ import android.widget.FrameLayout
 import android.widget.Toolbar
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
-import com.topjohnwu.superuser.Shell
 import nya.kitsunyan.foxydroid.R
-import nya.kitsunyan.foxydroid.content.Cache
 import nya.kitsunyan.foxydroid.content.Preferences
 import nya.kitsunyan.foxydroid.database.CursorOwner
 import nya.kitsunyan.foxydroid.utility.KParcelable
 import nya.kitsunyan.foxydroid.utility.Utils
-import nya.kitsunyan.foxydroid.utility.extension.android.*
+import nya.kitsunyan.foxydroid.utility.Utils.startPackageInstaller
 import nya.kitsunyan.foxydroid.utility.extension.resources.*
 import nya.kitsunyan.foxydroid.utility.extension.text.*
-import java.io.File
 
 abstract class ScreenActivity: FragmentActivity() {
   companion object {
@@ -210,7 +206,7 @@ abstract class ScreenActivity: FragmentActivity() {
           if (fragment !is ProductFragment || fragment.packageName != packageName) {
             pushFragment(ProductFragment(packageName))
           }
-          specialIntent.cacheFileName?.let(::startPackageInstaller)
+          specialIntent.cacheFileName?.let { startPackageInstaller(it) }
         }
         Unit
       }
@@ -230,37 +226,6 @@ abstract class ScreenActivity: FragmentActivity() {
       }
     }
   }
-
-  internal fun startPackageInstaller(cacheFileName: String) {
-    if (Preferences[Preferences.Key.RootInstalation] && Shell.getShell().isRoot) {
-      val commandBuilder = StringBuilder()
-      // disable verify apps over usb
-      commandBuilder.append("settings put global verifier_verify_adb_installs 0 ; ")
-      // Install main package
-      commandBuilder.append(
-        getPackageInstallCommand(Cache.getReleaseFile(this, cacheFileName))
-      )
-      // re-enable verify apps over usb after install
-      commandBuilder.append(" ; settings put global verifier_verify_adb_installs 1")
-      val result = Shell.su(commandBuilder.toString()).exec()
-    } else {
-      val (uri, flags) = if (Android.sdk(24)) {
-        Pair(Cache.getReleaseUri(this, cacheFileName), Intent.FLAG_GRANT_READ_URI_PERMISSION)
-      } else {
-        Pair(Uri.fromFile(Cache.getReleaseFile(this, cacheFileName)), 0)
-      }
-      // TODO Handle deprecation
-      @Suppress("DEPRECATION")
-      startActivity(
-        Intent(Intent.ACTION_INSTALL_PACKAGE)
-          .setDataAndType(uri, "application/vnd.android.package-archive").setFlags(flags)
-      )
-    }
-  }
-
-  private fun getPackageInstallCommand(apkPath: File): String =
-      "cat \"${apkPath.absolutePath}\" | pm install -t -r -S ${apkPath.length()}"
-
 
   internal fun navigateProduct(packageName: String) = pushFragment(ProductFragment(packageName))
   internal fun navigateRepositories() = pushFragment(RepositoriesFragment())
