+++ src/main/kotlin/nya/kitsunyan/foxydroid/utility/Utils.kt
@@ -1,17 +1,25 @@
 package nya.kitsunyan.foxydroid.utility
 
 import android.animation.ValueAnimator
+import android.app.Activity
 import android.content.Context
+import android.content.Intent
 import android.content.pm.Signature
 import android.content.res.Configuration
 import android.graphics.drawable.Drawable
+import android.net.Uri
 import android.os.LocaleList
 import android.provider.Settings
+import android.util.Log
+import com.topjohnwu.superuser.Shell
 import nya.kitsunyan.foxydroid.BuildConfig
 import nya.kitsunyan.foxydroid.R
+import nya.kitsunyan.foxydroid.content.Cache
+import nya.kitsunyan.foxydroid.content.Preferences
 import nya.kitsunyan.foxydroid.utility.extension.android.*
 import nya.kitsunyan.foxydroid.utility.extension.resources.*
 import nya.kitsunyan.foxydroid.utility.extension.text.*
+import java.io.File
 import java.security.MessageDigest
 import java.security.cert.Certificate
 import java.security.cert.CertificateEncodingException
@@ -96,5 +104,56 @@ object Utils {
     } else {
       Settings.Global.getFloat(context.contentResolver, Settings.Global.ANIMATOR_DURATION_SCALE, 1f) != 0f
     }
+  }
+
+  fun quote(string: String) = "\"${string.replace(Regex("""[\\$"`]""")) { c -> "\\${c.value}" }}\""
+
+  internal fun Activity.startPackageInstaller(cacheFileName: String) {
+    if (Preferences[Preferences.Key.RootInstallation] && Shell.getShell().isRoot) {
+      val releaseFile = Cache.getReleaseFile(this, cacheFileName)
+      val commandBuilder = StringBuilder()
+      // disable verify apps over usb
+      commandBuilder.append("settings put global verifier_verify_adb_installs 0 ; ")
+      // Install main package
+      commandBuilder.append(getPackageInstallCommand(releaseFile))
+      // re-enable verify apps over usb after install
+      commandBuilder.append(" ; settings put global verifier_verify_adb_installs 1")
+      val result = Shell.su(commandBuilder.toString()).exec()
+      if (result.isSuccess) Shell.su("${getUtilBoxPath()} rm ${quote(releaseFile.absolutePath)}")
+    } else {
+      val (uri, flags) = if (Android.sdk(24)) {
+        Pair(Cache.getReleaseUri(this, cacheFileName), Intent.FLAG_GRANT_READ_URI_PERMISSION)
+      } else {
+        Pair(Uri.fromFile(Cache.getReleaseFile(this, cacheFileName)), 0)
+      }
+      // TODO Handle deprecation
+      @Suppress("DEPRECATION")
+      startActivity(
+        Intent(Intent.ACTION_INSTALL_PACKAGE)
+          .setDataAndType(uri, "application/vnd.android.package-archive").setFlags(flags)
+      )
+    }
+  }
+
+  private fun getPackageInstallCommand(apkPath: File): String =
+    "cat \"${apkPath.absolutePath}\" | pm install -t -r -S ${apkPath.length()}"
+
+  private fun getUtilBoxPath(): String {
+    listOf("toybox", "busybox").forEach {
+      var shellResult = Shell.su("which $it").exec()
+      if (shellResult.out.isNotEmpty()) {
+        val utilBoxPath = shellResult.out.joinToString("")
+        if (utilBoxPath.isNotEmpty()) {
+          val utilBoxQuoted = quote(utilBoxPath)
+          shellResult = Shell.su("$utilBoxQuoted --version").exec()
+          if (shellResult.out.isNotEmpty()) {
+            val utilBoxVersion = shellResult.out.joinToString("")
+            Log.i(this.javaClass.canonicalName,"Using Utilbox $it : $utilBoxQuoted $utilBoxVersion")
+          }
+          return utilBoxQuoted
+        }
+      }
+    }
+    return ""
   }
 }
